requirements.txt
# Core Computer Vision and MediaPipe
opencv-python==4.8.1.78
mediapipe==0.10.21
numpy==1.24.3

# Input Control and Automation
pynput==1.7.6
pyautogui==0.9.54
pydirectinput==1.0.4

# Image Processing
Pillow==10.0.1

# Voice Recognition and Audio
SpeechRecognition==3.10.0
pyaudio==0.2.11
pyttsx3==2.90
fuzzywuzzy==0.18.0

# Machine Learning and AI
torch>=2.0.0
torchvision>=0.15.0
scikit-learn>=1.3.0
joblib>=1.3.0

# RAG System Dependencies (install in separate rag_env)
# chromadb>=0.4.0,<0.5.0
# langchain>=0.1.0
# openai>=1.0.0
# sentence-transformers>=2.2.0
# PyPDF2>=3.0.0
# pycryptodome>=3.0.0

# API and Web
requests==2.31.0
google-generativeai

# System Monitoring
psutil>=5.9.0

# Development and Testing
setuptools>=65.0.0
wheel>=0.37.0